requirements.txt
torch>=2.4.0
torchvision>=0.19.0
diffusers>=0.30.1
transformers>=4.44.0
accelerate>=0.33.0
sentencepiece>=0.2.0
SwissArmyTransformer>=0.4.12
numpy==1.26.0
gradio>=4.42.0
streamlit>=1.37.1
imageio==2.34.2
imageio-ffmpeg==0.5.1
openai>=1.42.0
moviepy==1.0.3
pillow==9.5.0

numpy
pytorch_lightning
omegaconf
einops
av==12.0.0
kornia
xformers==0.0.27.post2
open_clip_torch<=2.24.0
opencv_python
decord